+++ tests/integer_test.py
@@ -7,7 +7,7 @@ def test_fixed_byte_int():
     assert codec.encode(3) == Raw.hex("00 00 03")
     assert codec.decode(Raw.hex("01 00 02")) == 0x010002
 
-    codec = IntegerFormat(bytes=3, big_end=True).create_codec()
+    codec = IntegerFormat(bytes=3, lsb_first=True).create_codec()
     assert codec.encode(3) == Raw.hex("03 00 00")
     assert codec.decode(Raw.hex("01 00 02")) == 0x020001